Create part1.html
This creates the folder ‘templates’ and creates the file ‘login.html’, ‘register.html’, and ‘index.html’ inside the ‘templates’ folder. Then we open the ‘login.html’ file and write the code given below. In ‘login.html’, we have two fields i.e. username and password. When the user enters the correct username and password, it will route you to the index page otherwise ‘Incorrect username/password is displayed.