Create part4.css
We created the folder ‘static’ and a file ‘style.css’ inside the ‘static’ folder with the given CSS code. 
Run the server. Browse the URL ‘localhost:5000’. The output web page will be displayed.  Testing of the Application: If you are new user, go to sign up page and fill the details. After registration, go to login page. Enter your username and password and sign in. If your login is successful, you will be moved to index page and your name will be displayed.